Started implement plant DB since Trefle is discontinued.
Have yet to implement PlantBio.

+++ app/src/main/java/com/example/operationsunlight/modules/plant/OnPlantListener.java
@@ -1,5 +1,6 @@
 package com.example.operationsunlight.modules.plant;
 
 public interface OnPlantListener {
-    void onPlantClick(int plant_id);
+//    void onPlantClick(int plant_id);
+    void onPlantClick(long plant_id);
 }
